tried to use ReentrantReadWriteLocks to coordinate work spread out over ExecutorService (agents and other clojure based constructs don't map well to the dynamic nature of the task assignment in this case - we never really know at the start how many tasks there are so we can't just create the right number of agents for isntance).
These locks are not working as there is a race defect between when you
realize there is another task to add to the queue and when the locks are
acquired.

I think the only sure way to do this is to use an atom sort of like
a CyclicBarrier - we would watch it until it hits zero and then we are done.
Or maybe we use a clojure construct that allows for a "watcher" on it so
we can "push" the fact that the counter gets to zero.